+++ src/dsp_tools/commands/excel2xml/__init__.py
@@ -3,5 +3,5 @@
 
 from dsp_tools.commands.excel2xml.excel2xml_lib import *
 from dsp_tools.commands.excel2xml.propertyelement import PropertyElement as PropertyElement
-from dsp_tools.utils.shared import check_notna as check_notna
-from dsp_tools.utils.shared import simplify_name as simplify_name
+from dsp_tools.utils.date_formats.shared import check_notna as check_notna
+from dsp_tools.utils.date_formats.shared import simplify_name as simplify_name
