+++ src/dsp_tools/commands/project/create/project_create_lists.py
@@ -13,7 +13,7 @@ from dsp_tools.commands.project.legacy_models.listnode import ListNode
 from dsp_tools.commands.project.legacy_models.project import Project
 from dsp_tools.error.exceptions import BaseError
 from dsp_tools.error.exceptions import InputError
-from dsp_tools.utils.shared import parse_json_input
+from dsp_tools.utils.json_parsing import parse_json_input
 
 
 def create_lists_on_server(
