+++ src/dsp_tools/commands/xmlupload/prepare_xml_input/read_validate_xml_file.py
@@ -6,7 +6,7 @@ import regex
 from lxml import etree
 
 from dsp_tools.error.exceptions import InputError
-from dsp_tools.utils.iri_util import is_resource_iri
+from dsp_tools.utils.date_formats.iri_util import is_resource_iri
 from dsp_tools.utils.xml_parsing.parse_xml import parse_xml_file
 from dsp_tools.utils.xml_parsing.transform import remove_comments_from_element_tree
 from dsp_tools.utils.xml_parsing.transform import transform_special_tags_make_localname
