+++ src/dsp_tools/utils/date_formats/shared.py
@@ -1,18 +1,14 @@
 from __future__ import annotations
 
-import json
 import unicodedata
-from pathlib import Path
 from typing import Any
 from typing import Optional
 from typing import TypeGuard
-from typing import Union
 
 import pandas as pd
 import regex
 
 from dsp_tools.commands.excel2xml.propertyelement import PropertyElement
-from dsp_tools.error.exceptions import BaseError
 
 
 def simplify_name(value: str) -> str:
@@ -83,33 +79,3 @@ def check_notna(value: Optional[Any]) -> TypeGuard[Any]:
         )
     else:
         return False
-
-
-def parse_json_input(project_file_as_path_or_parsed: Union[str, Path, dict[str, Any]]) -> dict[str, Any]:
-    """
-    Check the input for a method that expects a JSON project definition, either as file path or as parsed JSON object:
-    If it is parsed already, return it unchanged.
-    If the input is a file path, parse it.
-
-    Args:
-        project_file_as_path_or_parsed: path to the JSON project definition, or parsed JSON object
-
-    Raises:
-        BaseError: if the input is invalid
-
-    Returns:
-        the parsed JSON object
-    """
-    project_definition: dict[str, Any] = {}
-    if isinstance(project_file_as_path_or_parsed, dict):
-        project_definition = project_file_as_path_or_parsed
-    elif isinstance(project_file_as_path_or_parsed, (str, Path)) and Path(project_file_as_path_or_parsed).exists():
-        with open(project_file_as_path_or_parsed, encoding="utf-8") as f:
-            try:
-                project_definition = json.load(f)
-            except json.JSONDecodeError as e:
-                msg = f"The input file '{project_file_as_path_or_parsed}' cannot be parsed to a JSON object."
-                raise BaseError(msg) from e
-    else:
-        raise BaseError("Invalid input: The input must be a path to a JSON file or a parsed JSON object.")
-    return project_definition
