+++ test/integration/commands/project/test_create_project.py
@@ -15,7 +15,7 @@ from dsp_tools.commands.project.create.project_validate import _identify_problem
 from dsp_tools.commands.project.create.project_validate import validate_project
 from dsp_tools.error.exceptions import BaseError
 from dsp_tools.error.exceptions import InputError
-from dsp_tools.utils.shared import parse_json_input
+from dsp_tools.utils.json_parsing import parse_json_input
 
 
 @pytest.fixture
